Added Service Worker for Push Notification

+++ server/src/main/java/ca/dal/comparify/alerts/AlertService.java
@@ -3,19 +3,25 @@ package ca.dal.comparify.alerts;
 import ca.dal.comparify.alerts.model.AlertModel;
 import ca.dal.comparify.alerts.model.AlertRequestModel;
 import ca.dal.comparify.alerts.model.AlertResponseModel;
-import ca.dal.comparify.framework.notification.model.AbstractChannelNotificationModel;
+import ca.dal.comparify.constant.MailTemplateConstant;
+import ca.dal.comparify.framework.notification.model.MailNotificationModel;
 import ca.dal.comparify.framework.notification.model.WebPushNotificationModel;
 import ca.dal.comparify.framework.notification.model.WebSocketNotificationModel;
 import ca.dal.comparify.model.HashModel;
 import ca.dal.comparify.notification.NotificationService;
 import ca.dal.comparify.notification.model.*;
+import ca.dal.comparify.user.model.iam.UserDetailsModel;
+import ca.dal.comparify.user.service.UserDetailsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.format.DateTimeFormatter;
+import java.time.format.FormatStyle;
 import java.util.Collections;
 import java.util.List;
 
 import static ca.dal.comparify.alerts.model.AlertTypeEnum.PRODUCT_INFORMATION_AVAILABLE;
+import static ca.dal.comparify.utils.DateUtils.zoneNow;
 import static ca.dal.comparify.utils.ObjectUtils.convert;
 
 /**
@@ -30,16 +36,32 @@ public class AlertService {
     @Autowired
     private NotificationService notificationService;
 
+    @Autowired
+    private UserDetailsService userDetailsService;
+
     /**
-     * @param model
+     * @param alert
      * @param createdBy
      * @return
      *
      * @author Harsh Shah
      */
-    public int create(AlertRequestModel model, String createdBy) {
-        return alertRepository.save(AlertModel.transform(model, createdBy));
+    public int create(AlertRequestModel alert, String createdBy) {
+        int status = alertRepository.save(AlertModel.transform(alert, createdBy));
+
+        if(status == 0){
+
+            String title = String.format("Successful Creation of Alert - %s", alert.getAlertIdentifier());
+            HashModel model = createMailModel(createdBy, alert.getAlertIdentifier());
+
+            notificationService.send(createdBy,
+                new MailNotificationModel(title, MailTemplateConstant.ALERT_UNSUBSCRIBE_TEMPLATE,
+                    IconType.ALERT, NotificationTypeEnum.ALERT, createdBy, model));
+        }
+
+        return status;
     }
+
 
     /**
      * @param userIdentifier
@@ -59,7 +81,26 @@ public class AlertService {
      * @author Harsh Shah
      */
     public boolean delete(String userId, String id) {
-        return alertRepository.delete(userId, id);
+
+        AlertModel alert = fetchAlertById(id);
+        if(alert == null){
+            return false;
+        }
+
+        boolean status = alertRepository.delete(userId, id);
+
+        if(status){
+
+            String title = String.format("Successful Deletion of Alert - %s", alert.getAlertIdentifier());
+            HashModel model = createMailModel(userId, alert.getAlertIdentifier());
+
+            notificationService.send(userId,
+                new MailNotificationModel(title, MailTemplateConstant.ALERT_UNSUBSCRIBE_TEMPLATE,
+                    IconType.ALERT, NotificationTypeEnum.ALERT, userId, model));
+        }
+
+        return status;
+
     }
 
     /**
@@ -140,6 +181,16 @@ public class AlertService {
     }
 
     /**
+     * @param alertId
+     * @return
+     *
+     * @author Harsh Shah
+     */
+    private AlertModel fetchAlertById(String alertId) {
+        return alertRepository.fetchAlertById(alertId);
+    }
+
+    /**
      * @param brandId
      * @param itemId
      * @return
@@ -148,5 +199,23 @@ public class AlertService {
      */
     private HashModel checkForAlerts(String brandId, String itemId) {
         return alertRepository.checkForAlerts(brandId, itemId);
+    }
+
+    /**
+     * @param userId
+     * @param alertIdentifier
+     * @return
+     *
+     * @author Harsh Shah
+     */
+    private HashModel createMailModel(String userId, String alertIdentifier){
+        UserDetailsModel user = userDetailsService.findUserById(userId);
+
+        HashModel model = new HashModel();
+        model.put("name", user.getFullName());
+        model.put("alertName", alertIdentifier);
+        model.put("subscriptionDate", zoneNow().format(DateTimeFormatter.ofLocalizedDate(FormatStyle.FULL)));
+
+        return model;
     }
 }
